# openapi.yaml
openapi: 3.1.0
info:
  title: Sistema Integral de Gestión de Consorcios
  description: |
    API REST para la gestión integral de consorcios. Permite administrar expensas,
    gastos del consorcio, personal y liquidaciones, tareas y presupuestos,
    comunicación interna, reserva de amenities y configuración global. El acceso a
    los recursos está restringido por roles: Administración, Departamentos
    (Inquilinos/Propietarios), Representantes y Super-Admin.
    Fase 3.5 (jun-2026): cuenta corriente por departamento — pagos se asignan FIFO desde movimientos.

    **Multitenant SaaS (Fase 7A, jul-2026)**: la plataforma es multi-tenant con
    jerarquía Super-Admin → Administración → Consorcio → Departamento. Todo
    endpoint operacional (fuera de `/auth/*` y `/me/*`) exige el header
    `X-Consorcio-Id: <id>` para resolver el consorcio activo. El backend valida
    que el usuario del JWT tenga acceso al consorcio pedido; caso contrario
    devuelve **403**. Códigos específicos: `403 cambio_password_requerido`
    (usuario con `must_change_password=True`), `403 administracion_suspendida`
    (login sobre administración inactiva), `403 sin acceso a este consorcio`
    (spoof cross-tenant), `400 X-Consorcio-Id requerido/invalido`.
  version: 1.1.0
servers:
  - url: http://localhost:8000
    description: Servidor local de desarrollo
...
security:
  - bearerAuth: []
paths:
...
  /amenities:
    get:
      tags: [Amenities]
      summary: Listar amenities
      description: |
        Devuelve los amenities disponibles en el consorcio (SUM, Laundry, etc.).
        Accesible para todos los roles autenticados.
      operationId: listarAmenities
      parameters:
        - name: incluir_inactivos
          in: query
          required: false
          description: Si true y el caller es admin, devuelve también inactivos. Para no-admin se ignora.
          schema: { type: boolean, default: false }
      responses:
        '200':
          description: Lista de amenities recuperada exitosamente.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Amenity'
        '401':
          $ref: '#/components/responses/NoAutenticado'

    post:
      tags: [Administracion]
      summary: Crear un amenity (espacio común)
      description: |
        Crea un nuevo amenity. Solo accesible para Administración. El `nombre`
        es único en el consorcio — si ya existe se devuelve 409.
      operationId: crearAmenity
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AmenityCrear'
      responses:
        '201':
          description: Amenity creado exitosamente.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Amenity'
        '400':
          $ref: '#/components/responses/PedidoInvalido'
        '401':
          $ref: '#/components/responses/NoAutenticado'
        '403':
          $ref: '#/components/responses/AccesoDenegado'
        '409':
          description: Ya existe un amenity con ese nombre.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /amenities/{amenity_id}:
    patch:
      tags: [Administracion]
      summary: Editar un amenity
      description: |
        Permite a Administración modificar `nombre` y/o `descripcion` de un
        amenity. Solo se actualizan los campos enviados (PATCH parcial). El
        nuevo `nombre` debe seguir siendo único (409 si colisiona con otro).
      operationId: actualizarAmenity
      parameters:
        - name: amenity_id
          in: path
          required: true
          description: Identificador único del amenity.
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AmenityActualizar'
      responses:
        '200':
          description: Amenity actualizado exitosamente.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Amenity'
        '400':
          $ref: '#/components/responses/PedidoInvalido'
        '401':
          $ref: '#/components/responses/NoAutenticado'
        '403':
          $ref: '#/components/responses/AccesoDenegado'
        '404':
          $ref: '#/components/responses/NoEncontrado'
        '409':
          description: Ya existe otro amenity con ese nombre.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      tags: [Amenities]
      summary: Dar de baja un amenity (soft-delete)
      operationId: darDeBajaAmenity
      parameters:
        - name: amenity_id
          in: path
          required: true
          schema: { type: integer, format: int64 }
      responses:
        '200':
          description: Amenity dado de baja exitosamente.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Amenity'
        '401':
          $ref: '#/components/responses/NoAutenticado'
        '403':
          $ref: '#/components/responses/AccesoDenegado'
        '404':
          $ref: '#/components/responses/NoEncontrado'
        '409':
          description: El amenity ya estaba inactivo.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /amenities/{amenity_id}/disponibilidad:
    get:
      tags: [Amenities]
      summary: Consultar disponibilidad de un amenity
      description: |
        Devuelve los bloques horarios disponibles y reservados para un amenity
        (SUM, Laundry, etc.) en un rango de fechas. Accesible para todos los roles.
      operationId: consultarDisponibilidad
      parameters:
        - name: amenity_id
          in: path
          required: true
          description: Identificador único del amenity.
          schema:
            type: integer
            format: int64
        - name: desde
          in: query
          required: true
          description: Fecha inicial del rango a consultar (ISO 8601).
          schema:
            type: string
            format: date
        - name: hasta
          in: query
          required: true
          description: Fecha final del rango a consultar (ISO 8601).
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Disponibilidad recuperada exitosamente.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Disponibilidad'
        '400':
          $ref: '#/components/responses/PedidoInvalido'
        '401':
          $ref: '#/components/responses/NoAutenticado'
        '403':
          $ref: '#/components/responses/AccesoDenegado'
        '404':
          $ref: '#/components/responses/NoEncontrado'
  /amenities/{amenity_id}/reservas:
    post:
      tags: [Amenities]
      summary: Reservar un amenity
      description: |
        Crea una nueva reserva para el amenity indicado. Accesible para todos
        los roles autenticados. El `usuario_id` se asigna server-side desde el
        token JWT; cualquier valor en el body es ignorado. `inicio` debe ser
        estrictamente anterior a `fin` (400 si no). Si el rango solicitado se
        superpone con una reserva confirmada existente para el mismo amenity,
        se devuelve 409 Conflict.
      operationId: crearReserva
      parameters:
        - name: amenity_id
          in: path
          required: true
          description: Identificador único del amenity.
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReservaCrear'
      responses:
        '201':
          description: Reserva creada exitosamente.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Reserva'
        '400':
          $ref: '#/components/responses/PedidoInvalido'
        '401':
          $ref: '#/components/responses/NoAutenticado'
        '403':
          $ref: '#/components/responses/AccesoDenegado'
        '404':
          $ref: '#/components/responses/NoEncontrado'
        '409':
          description: El horario solicitado se superpone con una reserva existente.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Autenticación mediante token JWT en el header Authorization.
...
  responses:
    PedidoInvalido:
      description: El pedido es inválido o le faltan campos requeridos.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NoAutenticado:
      description: No se proporcionó token de autenticación o es inválido.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    AccesoDenegado:
      description: El rol del usuario no tiene permisos para acceder al recurso.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NoEncontrado:
      description: El recurso solicitado no existe.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
    Error:
      type: object
      required: [detail]
      properties:
        detail:
          type: string
          description: Descripción legible del error.
          examples: ["No tiene permisos para acceder a este recurso."]

    # ---------------------------- Consorcios (Plan D) ---------------------
...
    Amenity:
      type: object
      required: [id, nombre, descripcion, activo, precio_reserva, duracion_maxima_horas, anticipacion_maxima_dias, max_reservas_activas_por_depto, horas_minimas_cancelacion]
      properties:
        id: { type: integer, format: int64 }
        nombre: { type: string }
        descripcion: { type: string, nullable: true }
        activo: { type: boolean }
        precio_reserva: { type: number, nullable: true }
        duracion_maxima_horas: { type: integer, nullable: true }
        anticipacion_maxima_dias: { type: integer, nullable: true }
        max_reservas_activas_por_depto: { type: integer, nullable: true }
        horas_minimas_cancelacion: { type: integer, nullable: true }
    AmenityCrear:
      type: object
      required: [nombre]
      properties:
        nombre: { type: string, minLength: 1, maxLength: 100 }
        descripcion: { type: string, maxLength: 500, nullable: true }
        precio_reserva: { type: number, minimum: 0, nullable: true }
        duracion_maxima_horas: { type: integer, exclusiveMinimum: 0, nullable: true }
        anticipacion_maxima_dias: { type: integer, exclusiveMinimum: 0, nullable: true }
        max_reservas_activas_por_depto: { type: integer, exclusiveMinimum: 0, nullable: true }
        horas_minimas_cancelacion: { type: integer, minimum: 0, nullable: true }
    AmenityActualizar:
      type: object
      properties:
        nombre: { type: string, minLength: 1, maxLength: 100, nullable: true }
        descripcion: { type: string, maxLength: 500, nullable: true }
        activo: { type: boolean, nullable: true }
        precio_reserva: { type: number, minimum: 0, nullable: true }
        duracion_maxima_horas: { type: integer, exclusiveMinimum: 0, nullable: true }
        anticipacion_maxima_dias: { type: integer, exclusiveMinimum: 0, nullable: true }
        max_reservas_activas_por_depto: { type: integer, exclusiveMinimum: 0, nullable: true }
        horas_minimas_cancelacion: { type: integer, minimum: 0, nullable: true }
...
    Disponibilidad:
      type: object
      required: [amenity_id, bloques]
      properties:
        amenity_id:
          type: integer
          format: int64
        bloques:
          type: array
          items:
            $ref: '#/components/schemas/BloqueHorario'
    BloqueHorario:
      type: object
      required: [inicio, fin, disponible]
      properties:
        inicio:
          type: string
          format: date-time
        fin:
          type: string
          format: date-time
        disponible:
          type: boolean
    Reserva:
      type: object
      required: [id, amenity_id, usuario_id, inicio, fin, estado, movimiento_cuenta_id]
      properties:
        id: { type: integer, format: int64 }
        amenity_id: { type: integer, format: int64 }
        usuario_id: { type: integer, format: int64 }
        inicio: { type: string, format: date-time }
        fin: { type: string, format: date-time }
        estado: { type: string, enum: [confirmada, cancelada] }
        movimiento_cuenta_id: { type: integer, format: int64, nullable: true }
    ReservaCrear:
      type: object
      required: [inicio, fin]
      properties:
        inicio:
          type: string
          format: date-time
        fin:
          type: string
          format: date-time
...